chore(client): rename morphoto-header
