--- client/src/app/_component/mor-photo-header/index.ts
@@ -1,1 +1,0 @@
-export * from "./mor-photo-header";
