+++ client/src/app/_component/morphoto-header/index.ts
@@ -1,0 +1,1 @@
+export * from "./morphoto-header";
