+++ client/src/app/layout.tsx
@@ -1,7 +1,7 @@
 import "./globals.css";
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
-import { MorPhotoHeader } from "./_component";
+import { MorphotoHeader } from "./_component";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -18,7 +18,7 @@ export default function RootLayout({
   return (
     <html lang="en">
       <body className={inter.className}>
-        <MorPhotoHeader />
+        <MorphotoHeader />
         <div
           style={{
             paddingTop: 52,
